requirements.txt
asgiref==3.8.1
captcha==0.6.0
crispy-bootstrap5==2024.10
Django==5.1.3
django-allauth==65.3.0
django-captcha==1.0.0
django-crispy-forms==2.3
django-recaptcha==4.0.0
django-tinymce==4.1.0
gunicorn==23.0.0
loadenv==0.1.1
packaging==24.2
pillow==11.0.0
psycopg2==2.9.10
python-dotenv==1.0.1
recaptcha==1.0rc1
six==1.17.0
sqlparse==0.5.2
whitenoise==6.8.2
django-phone-field==1.8.1
django-phonenumber-field==8.0.0
phonenumbers==8.13.52